feat: dedicated scan queue page with nav badge
The queue was only reachable by opening the scanner modal from Search, so a
batch recognizing in the background was effectively invisible — which defeats
the point of queueing it in the first place.

- New /scans route (and /scans/:jobId) listing jobs with progress and status,
  drilling into the per-item review list.
- Review UI moved out of the scanner modal onto the page. The modal is now
  capture-only: queueing a batch closes it and navigates to the new job.
  Single-photo scans still answer inline in the modal, unchanged.
- Shared review components extracted to components/ScanReview.jsx so the modal
  (single scan) and the page (queued batches) present the same add-to-collection
  step instead of duplicating it.
- Sidebar and mobile nav show a scan entry with an outstanding-item badge only
  when there is something running or to review, matching the existing
  conditional migration-link pattern. Spinner while work is in flight.
- job_progress now returns `unresolved` so the badge counts what actually needs
  attention — reviewing an item decrements it and failed items don't keep it
  lit, which client-side math from done/pending could not get right.
- Polling backs off to 60s when nothing is active instead of running at 5s
  forever.

Frontend still unverified by vite/vitest (no node in this environment):
manual review, import/usage audit, and a brace-balance check only.
